fix: move all element ids to uint32

+++ contracts/src/interfaces/IRedeemManager.1.sol
@@ -14,14 +14,14 @@ interface IRedeemManagerV1 {
     /// @param height The height of the redeem request in LsETH
     /// @param size The size of the redeem request in LsETH
     /// @param id The id of the new redeem request
-    event CreatedRedeemRequest(address indexed owner, uint256 height, uint256 size, uint256 id);
+    event CreatedRedeemRequest(address indexed owner, uint256 height, uint256 size, uint32 id);
 
     /// @notice Emitted when a withdrawal event is created
     /// @param height The height of the withdrawal event in LsETH
     /// @param size The size of the withdrawal event in LsETH
     /// @param ethAmount The amount of eth to distrubute to claimers
     /// @param id The id of the withdrawal event
-    event CreatedWithdrawalEvent(uint256 height, uint256 size, uint256 ethAmount, uint256 id);
+    event CreatedWithdrawalEvent(uint256 height, uint256 size, uint256 ethAmount, uint32 id);
 
     /// @notice Emitted when a redeem request has been filled (even partially)
     /// @param id The id of the redeem request
@@ -29,7 +29,7 @@ interface IRedeemManagerV1 {
     /// @param amountFilled The amount of LsETH filled
     /// @param ethAmountFilled The amount of ETH filled
     event FilledRedeemRequest(
-        uint256 indexed id, uint256 withdrawalEventId, uint256 amountFilled, uint256 ethAmountFilled
+        uint32 indexed id, uint32 withdrawalEventId, uint256 amountFilled, uint256 ethAmountFilled
     );
 
     /// @notice Emitted when a account is sent rewards
@@ -76,7 +76,7 @@ interface IRedeemManagerV1 {
     /// @notice Retrieve the details of a specific redeem request
     /// @param redeemRequestId The id of the request
     /// @return The redeem request details
-    function getRedeemRequestDetails(uint256 redeemRequestId)
+    function getRedeemRequestDetails(uint32 redeemRequestId)
         external
         view
         returns (RedeemRequests.RedeemRequest memory);
@@ -87,7 +87,7 @@ interface IRedeemManagerV1 {
     /// @notice Retrieve the details of a specific withdrawal event
     /// @param withdrawalEventId The id of the withdrawal event
     /// @return The withdrawal event details
-    function getWithdrawalEventDetails(uint256 withdrawalEventId)
+    function getWithdrawalEventDetails(uint32 withdrawalEventId)
         external
         view
         returns (WithdrawalEvents.WithdrawalEvent memory);
@@ -95,7 +95,7 @@ interface IRedeemManagerV1 {
     /// @notice Retrieve the list of redeem requests of an account
     /// @param account The account to query
     /// @return redeemRequestIds The list of redeemRequests belonging to the specified account
-    function listRedeemRequests(address account) external view returns (uint256[] memory redeemRequestIds);
+    function listRedeemRequests(address account) external view returns (uint32[] memory redeemRequestIds);
 
     /// @notice Resolves the provided list of redeem request ids
     /// @dev The result is an array of equal length with ids or error code
@@ -105,24 +105,24 @@ interface IRedeemManagerV1 {
     /// @dev This call was created to be called by an off-chain interface, the output could then be used to perform the claimRewards call in a regular transaction
     /// @param redeemRequestIds The list of redeem requests to resolve
     /// @return withdrawalEventIds The list of withdrawal events matching every redeem request (or error codes)
-    function resolveRedeemRequests(uint256[] calldata redeemRequestIds)
+    function resolveRedeemRequests(uint32[] calldata redeemRequestIds)
         external
         view
-        returns (int256[] memory withdrawalEventIds);
+        returns (int64[] memory withdrawalEventIds);
 
     /// @notice Creates a redeem request
     /// @param lsETHAmount The amount of LsETH to redeem
     /// @param recipient The recipient owning the redeem request
     /// @return redeemRequestId The id of the redeem request
-    function requestRedeem(uint256 lsETHAmount, address recipient) external returns (uint256 redeemRequestId);
+    function requestRedeem(uint256 lsETHAmount, address recipient) external returns (uint32 redeemRequestId);
 
     /// @notice Claims the rewards of the provided redeem request ids
     /// @param redeemRequestIds The list of redeem requests to claim
     /// @param withdrawalEventIds The list of withdrawal events to use for every redeem request claim
     /// @param skipAlreadyClaimed True if the call should not revert on claiming of already claimed requests
     function claimRewards(
-        uint256[] calldata redeemRequestIds,
-        uint256[] calldata withdrawalEventIds,
+        uint32[] calldata redeemRequestIds,
+        uint32[] calldata withdrawalEventIds,
         bool skipAlreadyClaimed
     ) external;
